requirements.txt
streamlit
pandas
pytesseract
openpyxl
pillow
opencv-python-headless
